add token check

+++ .bin/render/render-put-target-repository-clone.sh
@@ -6,6 +6,11 @@ TARGET_REPO_OWNER=${TARGET_REPO_OWNER:-$(${SCRIPTS_DIR}/render/render-get-source
 TARGET_REPO_NAME=${TARGET_REPO_NAME:-flux-platform-rendered}
 GH_TOKEN=${RENDER_GITHUB_TOKEN:-$(gh auth token)}
 
+echo "Cloning Target Repository: ${TARGET_REPO_OWNER}/${TARGET_REPO_NAME}"
+if [ -z "$GH_TOKEN" ]; then
+  echo "Error: GH_TOKEN is not set. Please set the RENDER_GITHUB_TOKEN environment variable or ensure you are authenticated with gh."
+  exit 1
+fi
 pushd ${RENDER_DIR} > /dev/null || exit 1
 rm -rf ${TARGET_REPO_NAME}
 gh repo clone ${TARGET_REPO_OWNER}/${TARGET_REPO_NAME} -- --depth 1
